feat(votos): añade empate neutral (#195)
* feat(votos): añade empate neutral

* fix(votos): soporta empate neutral en similares

* fix(votos): reparte empate en stats

+++ backend/src/main/java/com/diegoalegil/animeshowdown/dto/RankingMovimientoItem.java
@@ -15,7 +15,7 @@ public record RankingMovimientoItem(
         String nombre,
         String anime,
         String imagenUrl,
-        long votos,
+        double votos,
         int posicionActual,
         Integer posicionAnterior,
         Integer delta,
